test(client): colocate tests with source files
Move tests/brand-style-to-theme.test.ts and tests/find-template-owner.test.ts
into src/ next to the files they test. The vitest config already scans
src/ only, so the tests/ copies were never running.

- Merge three additional test cases from tests/brand-style-to-theme.test.ts
  (unsafe social link, font URL, and logo URL filtering) into the src/ copy.
- Fix one incorrect assertion in the unsafe-social-links test: @braintree/sanitize-url
  passes protocol-less strings (e.g. 'not-a-url') through unchanged; applyTheme
  does not reject them.
- Update find-template-owner imports from ../src/X.js to ./X.js.
